Update helper script to work on Unix and Osx

+++ helpers/replace_with_env_vars_values.sh
@@ -7,9 +7,23 @@ envfile=$1
 dstfile=$2
 
 echo "[TERRAFORM] Loading environment from '${envfile}'"
-for key in $( cat ${envfile} | awk '{$1=$1};1' | egrep -o "TF_VAR_[_A-Za-z]+" ) ; do
-    source ${envfile} ; echo -e "\t[TERRAFORM] Injecting '${key}=${!key}'" ;
-    source ${envfile} ; sed -E -i ".bak" "s/${key}(\W|$)/\"${!key}\"/g" ${dstfile} ;
-done
-echo "[HOUSEKEEPING] Clean unneeded backup file '${dstfile}"
-rm -f ${dstfile}.bak
+
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    for key in $( cat ${envfile} | awk '{$1=$1};1' | egrep -o "TF_VAR_[_A-Za-z]+" ) ; do
+        source ${envfile} ; echo -e "\t[TERRAFORM] Injecting '${key}=${!key}'" ;
+        source ${envfile} ; sed -r -i "s/${key}(\W|$)/\"${!key}\"/g" ${dstfile} ;
+    done
+
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    for key in $( cat ${envfile} | awk '{$1=$1};1' | egrep -o "TF_VAR_[_A-Za-z]+" ) ; do
+        source ${envfile} ; echo -e "\t[TERRAFORM] Injecting '${key}=${!key}'" ;
+        source ${envfile} ; sed -E -i ".bak" "s/${key}(\W|$)/\"${!key}\"/g" ${dstfile} ;
+    done
+    echo "[HOUSEKEEPING] Clean unneeded backup file '${dstfile}"
+    rm -f ${dstfile}.bak
+
+else
+    echo "Unsupported OS: please use Mac or Linux"
+fi
+
+
